requirements.txt
requests
python-decouple
sigma-cli
pyyaml
